pyproject.toml
[project]
name = "garmin-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.15",
    "garminconnect>=0.2.28",
    "openai>=1.97.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "typer>=0.16.0",
]
